Moved i18n module to commons package
